+++ src/addons/system/system.h
@@ -24,10 +24,10 @@ typedef struct ecs_system_t {
     bool multi_threaded;
     bool no_staging;
 
-    int32_t invoke_count;           /* Number of times system is invoked */
+    int64_t invoke_count;           /* Number of times system is invoked */
     float time_spent;               /* Time spent on running system */
     ecs_ftime_t time_passed;        /* Time passed since last invocation */
-    int32_t last_frame;             /* Last frame for which the system was considered */
+    int64_t last_frame;             /* Last frame for which the system was considered */
 
     void *ctx;                      /* Userdata for system */
     void *binding_ctx;              /* Optional language binding context */
